Have added some example cases of valid and invalid IPv6 addresses. These will be used as tests once complete. Have created a general skeleton that should pass standard IPv6 without IPv4 embedding however, has not been tested yet. This commit is just to put down initial thoughts. I'm sure the code isn't that tidy at the moment however, this isn't the finished produce. For now, I just need *something* to be written down which works and then that can be improved on later
Little more detail on how the code works:
When a double colon appears for example, "123:23:323::12" then .split(":") should cause one of the blocks to be empty, ["123","23","323","","12"]
The complexity comes from the fact that when you have the double colon trailing or leading  then an extra empty block is created for example, "x::" or "::x" makes ["x","",""] or ["","","x"] (respectively)
In valid cases this only happens if the 2 empty blocks are next to each other. So we have a counter for number of empty blocks we've seen, `number_of_empty`, and once we've seen a second empty block then we check if it came straight after the first one.

During commit thought:
	I've just realised that for example "x1::x2:x3::x4" would return True when should be seen as invalid. Perhaps need another bit of logic to say if `number_of_empty` exceeds 1 then we return False.